Avoid hard encoding and move TSStatusType.java to service-rpc
