+++ jdbc/src/test/java/org/apache/iotdb/jdbc/IoTDBStatementTest.java
@@ -31,6 +31,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
+import org.apache.iotdb.rpc.TSStatusType;
 import org.apache.iotdb.service.rpc.thrift.*;
 import org.apache.iotdb.service.rpc.thrift.TSIService.Iface;
 import org.apache.thrift.TException;
@@ -55,7 +56,7 @@ public class IoTDBStatementTest {
   @Mock
   private TSFetchMetadataResp fetchMetadataResp;
 
-  private TS_StatusType successStatus = new TS_StatusType(200, "");
+  private TS_StatusType successStatus = new TS_StatusType(TSStatusType.SUCCESS_STATUS.getStatusCode(), "");
   private TS_Status Status_SUCCESS = new TS_Status(successStatus);
   private ZoneId zoneID = ZoneId.systemDefault();
